+++ apps/firebase-chat/src/main.ts
@@ -24,17 +24,16 @@ import App from './App.vue';
 
 const app = createApp(App);
 
+const pinia = createPinia();
+pinia.use(piniaPluginPersistedState);
+app.use(pinia);
+
 const routes = setupLayouts(generatedRoutes);
 const router = createRouter({
   history: createWebHistory(),
   routes,
 });
 app.use(router);
-
-const pinia = createPinia();
-pinia.use(piniaPluginPersistedState);
-app.use(pinia);
-
 
 app.use(VueFire, {
   firebaseApp,
